Update README.md
edited README to add information about the classifier/sentiment analysis. The info about the second model with feature extraction is not added yet. Additionally, we need to add training error to the word2vec classification notebook (maybe do a classification report maybe a few extra lines of code). Also the conclusion part and the 'potential next models par needs to be filled in/updated'